+++ src/cli/chain/list/index.js
@@ -9,5 +9,7 @@ cli('chain:ls', 'List all chains', input => action(execute, run, input), {
     rpcClient: false,
     runningNode: false,
     wallet: false,
+    chainName: false,
+    chainExists: false,
   },
 });