+++ src/cli/misc/help/index.js
@@ -8,6 +8,8 @@ cli('help <subcommand>', 'Show help of a sub command', input => action(execute, 
     runningNode: false,
     rpcClient: false,
     wallet: false,
+    chainName: false,
+    chainExists: false,
   },
   options: [],
 });