+++ src/cli/misc/version/index.js
@@ -2,11 +2,18 @@
 const { cli, action } = require('core/cli');
 const { execute, run } = require('./version');
 
-cli('version', 'Output version for all forge components', input => action(execute, run, input), {
-  requirements: {
-    forgeRelease: true,
-    rpcClient: true,
-    wallet: false,
-  },
-  options: [],
-});
+cli(
+  'version [<chainName>]',
+  'Output version for all forge components',
+  input => action(execute, run, input),
+  {
+    requirements: {
+      forgeRelease: true,
+      rpcClient: true,
+      wallet: false,
+      chainName: 'required',
+      chainExists: true,
+    },
+    parseArgs: chainName => ({ chainName }),
+  }
+);