+++ src/cli/release/download/index.js
@@ -13,6 +13,8 @@ cli(
       runningNode: false,
       rpcClient: false,
       wallet: false,
+      chainName: false,
+      chainExists: false,
     },
     options: [['-f, --force', 'Clean local downloaded assets before download']],
     handlers: {