+++ src/cli/release/remote/index.js
@@ -12,6 +12,8 @@ cli(
       runningNode: false,
       rpcClient: false,
       wallet: false,
+      chainName: false,
+      chainExists: false,
     },
     options: [],
   }