+++ src/cli/release/use/index.js
@@ -1,7 +1,10 @@
 const shell = require('shelljs');
 const chalk = require('chalk');
 const { cli, action } = require('core/cli');
+const { getMinSupportForgeVersion } = require('core/libs/common');
 const { execute, run } = require('./use');
+
+const minSupportVersion = getMinSupportForgeVersion();
 
 cli(
   'use [version]',
@@ -13,14 +16,16 @@ cli(
       runningNode: false,
       rpcClient: false,
       wallet: false,
+      chainName: false,
+      chainExists: false,
     },
     options: [],
     handlers: {
       '--help': () => {
         shell.echo(`
 Examples:
-  - ${chalk.cyan('forge use 0.22.0')}      activate forge v0.22.0
-  - ${chalk.cyan('forge use v0.22.0')}     activate forge v0.22.0
+  - ${chalk.cyan(`forge use ${minSupportVersion}`)}      activate forge v${minSupportVersion}
+  - ${chalk.cyan(`forge use v${minSupportVersion}`)}     activate forge v${minSupportVersion}
         `);
       },
     },