+++ src/cli/release/use/use.js
@@ -1,11 +1,12 @@
 const chalk = require('chalk');
-const shell = require('shelljs');
 const semver = require('semver');
-const { symbols } = require('core/ui');
+
 const { isForgeStarted } = require('core/forge-process');
 const debug = require('core/debug')('release:use');
 const { updateReleaseYaml, listReleases, getGlobalForgeVersion } = require('core/forge-fs');
-const { print, printError, printSuccess } = require('core/util');
+const { print, printError, printSuccess, printWarning } = require('core/util');
+
+const { version: cliVersion, engines } = require('../../../../package.json');
 
 // eslint-disable-next-line consistent-return
 async function main({
@@ -20,25 +21,31 @@ async function main({
       process.exit(1);
     }
 
+    if (!semver.satisfies(userVersion, engines.forge)) {
+      printError(
+        `forge-cli@${cliVersion} requires forge@${engines.forge} to work, but got ${userVersion}!`
+      );
+      process.exit(1);
+    }
+
     const { version } = semver.coerce(userVersion);
-    if (semver.eq(version, getGlobalForgeVersion())) {
-      shell.echo(`${symbols.warning} Already using forge release v${version}`);
-      return process.exit(1);
+    const globalVersion = getGlobalForgeVersion();
+    if (semver.valid(globalVersion) && semver.eq(version, globalVersion)) {
+      printWarning(`Already using forge release v${version}`);
+      return process.exit(0);
     }
 
     if (allowMultiChain === false) {
       if (await isForgeStarted()) {
-        shell.echo(`${symbols.warning} Please stop forge before activate another version`);
+        printWarning('Please stop forge before activate another version');
         return process.exit(1);
       }
     }
 
     const releases = await listReleases();
     if (!releases.find(item => semver.eq(version, item.version))) {
-      shell.echo(
-        `${
-          symbols.error
-        } forge release v${version} not downloaded, please download it with ${chalk.cyan(
+      printError(
+        `forge release v${version} not downloaded, please download it with ${chalk.cyan(
           `forge download ${version}`
         )}`
       );
@@ -50,9 +57,10 @@ async function main({
     debug(`'${chainName}' chain version updated:`, version);
 
     printSuccess(`forge v${version} activated successfully!`);
-    print('');
-    print(`Now you can start forge with ${chalk.cyan(`forge start ${chainName}`)}`);
-    print('');
+    if (chainName) {
+      print();
+      print(`Now you can start forge with ${chalk.cyan(`forge start ${chainName}`)}`);
+    }
   } catch (err) {
     printError(err);
     printError('Forge release activate failed');