+++ src/core/cli.js
@@ -3,6 +3,7 @@ const path = require('path');
 const last = require('lodash/last');
 const rcfile = require('rcfile');
 const registryUrl = require('registry-url');
+const pickBy = require('lodash/pickBy');
 
 const { setupEnv } = require('./env');
 
@@ -15,14 +16,16 @@ const globalConfig = rcfile('forge');
  * @param {String} command command line
  * @param {String} desc documentation of the cli
  * @param {Function} handler cli handler
+ * @param {Function} parseArg command's arg parser
  */
 function cli(
   command,
   desc,
   handler,
-  { requirements = {}, options = [], alias, handlers = {} } = {}
+  { requirements = {}, options = [], alias, handlers = {}, parseArgs } = {}
 ) {
   allCommands.push({
+    parseArgs,
     command,
     desc,
     handler,
@@ -93,8 +96,17 @@ function initCli(program) {
           globalOpts.autoUpgrade = true;
         }
 
-        const opts = Object.assign({}, globalOpts, command.opts());
-        await setupEnv(globalArgs.args, x.requirements, opts);
+        const opts = Object.assign(
+          {},
+          pickBy(globalOpts, k => k !== undefined),
+          pickBy(command.opts(), k => k !== undefined)
+        );
+
+        if (typeof x.parseArgs === 'function') {
+          Object.assign(opts, pickBy(x.parseArgs(...params) || {}), k => k !== undefined);
+        }
+
+        await setupEnv(x.requirements, opts);
         await x.handler({
           args: params.filter(p => typeof p === 'string'),
           opts,